refactor: another pedantic change to public state naming (#1359)
# Description

I was writing docs, and something still felt a bit off when trying to
explain and justify the process of `PublicState` initialisation.
Renaming things helped describe what was happening.

I renamed structs from `MyTypeSerialisationInterface` to
`MyTypeSerialisationMethods`. Only `TypeSerialisationInterface` is an
interface. We then create global instances of methods which adhere to
this interface.

Similarly, I've renamed `MyNoteInterface` structs to be `MyNoteMethods`,
because they're collections of methods which adhere to the
`NoteInterface`. E.g. `ValueNoteMethods` implement the `NoteInterface`.