docs: Dapp tutorial (#2109)
Tutorial on building a dapp. I went for the simplest possible options in
order to focus on aztec.js usage and avoid the need for other prior
knowledge: I used javascript instead of typescript, and set up a console
app instead of a web one. Does not include anything on account creation
since that is not a dapp concern but a wallet's responsibility.

Pending:
- [x] Add links to the next section on each page of the tutorial
- [x] Add a small page on testing (with a link to the longer testing
guide)
- [x] Run the code on the CI

Fixes #2047 
Fixes #2048 
Fixes #2049 
Fixes #2050 
Fixes #2051